+++ frontend/src/pages/_app.tsx
@@ -1,3 +1,7 @@
+import { ThemeProvider as EmotionThemeProvider } from "@emotion/react";
+import { StylesProvider, ThemeProvider } from "@material-ui/core/styles";
+import createTheme from "@material-ui/core/styles/createTheme";
+import { defaultAppTheme, makeThemeOptions } from "czifui";
 import { NextPage } from "next";
 import { AppProps } from "next/app";
 import Script from "next/script";
@@ -23,15 +27,53 @@ type AppPropsWithLayout = AppProps & {
   Component: NextPageWithLayout;
 };
 
+const customTheme = {
+  typography: {
+    ...defaultAppTheme.typography,
+    fontFamily: "Roboto",
+  },
+};
+
+const themeOptions = { ...defaultAppTheme, ...customTheme };
+
+const primaryColors = {
+  "100": "#dfefff",
+  "200": "#73b7ff",
+  "300": "#459fff",
+  "400": "#0e7dec",
+  "500": "#005fc6",
+  "600": "#004c9f",
+};
+
+const infoColors = {
+  "100": "#73b7ff",
+  "200": "#F3EDFC",
+  "400": "#0e7dec",
+  "600": "#004c9f",
+};
+
+themeOptions.colors.primary = primaryColors;
+themeOptions.colors.info = infoColors;
+
+const appTheme = makeThemeOptions(themeOptions);
+
+const theme = createTheme(appTheme);
+
 function App({ Component, pageProps }: AppPropsWithLayout): JSX.Element {
   const Layout = Component.Layout || DefaultLayout;
   return (
     <>
       <QueryClientProvider client={queryClient}>
-        <Layout>
-          <Component {...pageProps} />
-        </Layout>
-        <ReactQueryDevtools />
+        <StylesProvider injectFirst>
+          <EmotionThemeProvider theme={theme}>
+            <ThemeProvider theme={theme}>
+              <Layout>
+                <Component {...pageProps} />
+              </Layout>
+              <ReactQueryDevtools />
+            </ThemeProvider>
+          </EmotionThemeProvider>
+        </StylesProvider>
       </QueryClientProvider>
       <Script
         data-domain={configs.PLAUSIBLE_DATA_DOMAIN}
